wip: Rest of sdf-server-v2 extraction
