+++ lib/axum-util/src/service.rs
@@ -9,6 +9,7 @@ use telemetry::prelude::*;
 use tracing_tunnel::TracingLevel;
 
 pub mod force_change_set_response;
+pub mod variant;
 pub mod ws;
 
 #[derive(Debug, Serialize)]
