+++ lib/sdf-server-v2-variant/src/create_unlocked_copy.rs
@@ -7,7 +7,7 @@ use dal::{
     SchemaVariant, SchemaVariantId, WorkspacePk, WsEvent,
 };
 
-use crate::{
+use axum_util::{
     extract::{AccessBuilder, HandlerContext, PosthogClient},
     service::{
         force_change_set_response::ForceChangeSetResponse,
