+++ lib/sdf-server-v2-variant/src/generate_aws_asset_schema.rs
@@ -6,7 +6,7 @@ use dal::{
 };
 use serde::{Deserialize, Serialize};
 
-use crate::{
+use axum_util::{
     extract::{AccessBuilder, AssetSprayer, HandlerContext, PosthogClient},
     service::force_change_set_response::ForceChangeSetResponse,
     track,
