+++ lib/sdf-server-v2-variant/src/get_variant.rs
@@ -5,9 +5,9 @@ use axum::{
 use dal::{ChangeSetId, SchemaVariant, SchemaVariantId, WorkspacePk};
 use si_frontend_types as frontend_types;
 
-use crate::{
+use crate::SchemaVariantsAPIError;
+use axum_util::{
     extract::{AccessBuilder, HandlerContext, PosthogClient},
-    service::v2::variant::SchemaVariantsAPIError,
     track,
 };
