+++ lib/sdf-server-v2-view/Cargo.toml
@@ -7,3 +7,13 @@ repository.workspace = true
 edition.workspace = true
 rust-version.workspace = true
 publish.workspace = true
+
+[dependencies]
+axum-util = { path = "../../lib/axum-util" }
+dal = { path = "../../lib/dal" }
+
+axum = { workspace = true }
+thiserror = { workspace = true }
+serde = { workspace = true }
+remain = { workspace = true }
+serde_json = { workspace = true }
