+++ lib/sdf-server-v2-view/src/update_view.rs
@@ -1,11 +1,17 @@
-use crate::extract::{AccessBuilder, HandlerContext, PosthogClient};
-use crate::service::force_change_set_response::ForceChangeSetResponse;
-use crate::service::v2::view::{ViewError, ViewResult};
-use crate::tracking::track;
-use axum::extract::{Host, OriginalUri, Path};
-use axum::Json;
-use dal::diagram::view::{View, ViewId};
-use dal::{ChangeSet, ChangeSetId, WorkspacePk};
+use crate::{ViewError, ViewResult};
+use axum::{
+    extract::{Host, OriginalUri, Path},
+    Json,
+};
+use axum_util::{
+    extract::{AccessBuilder, HandlerContext, PosthogClient},
+    service::force_change_set_response::ForceChangeSetResponse,
+    tracking::track,
+};
+use dal::{
+    diagram::view::{View, ViewId},
+    ChangeSet, ChangeSetId, WorkspacePk,
+};
 use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug)]
