+++ lib/sdf-server-v2/Cargo.toml
@@ -9,10 +9,14 @@ rust-version.workspace = true
 publish.workspace = true
 
 [dependencies]
+axum-util = { path = "../../lib/axum-util" }
 sdf-server-v2-admin = { path = "../../lib/sdf-server-v2-admin" }
 sdf-server-v2-audit-log = { path = "../../lib/sdf-server-v2-audit-log" }
+sdf-server-v2-change-set = { path = "../../lib/sdf-server-v2-change-set" }
 sdf-server-v2-func = { path = "../../lib/sdf-server-v2-func" }
 sdf-server-v2-management = { path = "../../lib/sdf-server-v2-management" }
 sdf-server-v2-module = { path = "../../lib/sdf-server-v2-module" }
 sdf-server-v2-variant = { path = "../../lib/sdf-server-v2-variant" }
 sdf-server-v2-view = { path = "../../lib/sdf-server-v2-view" }
+
+axum = { workspace = true }
