+++ lib/sdf-server/src/routes.rs
@@ -67,7 +67,7 @@ pub fn routes(state: AppState) -> Router {
         .nest("/api/ws", crate::service::ws::routes())
         .nest("/api/module", crate::service::module::routes())
         .nest("/api/variant", crate::service::variant::routes())
-        .nest("/api/v2", crate::service::v2::routes(state.clone()))
+        .nest("/api/v2", sdf_server_v2::routes(state.clone()))
         .layer(CompressionLayer::new())
         // allows us to be permissive about cors from our owned subdomains
         .layer(
